user auth

+++ app/Http/Controllers/ListingController.php
@@ -42,6 +42,8 @@ class ListingController extends Controller
             $formFields['logo'] = $request->file('logo') ->store('logos', 'public');
         }
 
+        $formFields['user_id'] = auth()->id();
+
         Listing::create($formFields);
 
         return redirect('/')->with('message', 'Listing created successfully!');
@@ -52,6 +54,10 @@ class ListingController extends Controller
     }
     //Update Listing Data
     public function update(Request $request, Listing $listing){
+        //validate logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
         $formFields = $request ->validate([
             'title' => 'required',
             'company' => ['required'],
@@ -72,7 +78,16 @@ class ListingController extends Controller
     }
     //Delete listing
     public function delete(Listing $listing) {
+        if($listing->user_id != auth()->id()){
+            abort(403, 'Unauthorized Action');
+        }
+        
         $listing->delete();
         return redirect('/')->with('message', 'Listing deleted successfully');
+    }
+    
+    //manage listings
+    public function manage(){
+        return view('listings.manage', ['listings'=> auth()->user()->listings()->get()]);
     }
 }
