+++ routes/web.php
@@ -3,6 +3,7 @@
 use App\Models\Listing;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
+use App\Http\Controllers\UserController;
 use App\Http\Controllers\ListingController;
 /*
 |--------------------------------------------------------------------------
@@ -27,22 +28,41 @@ use App\Http\Controllers\ListingController;
 Route::get('/', [ListingController::class, 'index']);
 
 //Show Create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 //Store listing Data
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
 //Show edit Form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 //Update Listing
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 //Delete Listing
-Route::delete('/listings/{listing}', [ListingController::class, 'delete']);
+Route::delete('/listings/{listing}', [ListingController::class, 'delete'])->middleware('auth');
+
+//Manage listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 //Single listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
+
+//Show Register/Create form
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
+
+//Create New User
+Route::post('/users', [UserController::class, 'store']);
+
+//Log User out
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
+
+//Show login form
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
+
+//Login user
+Route::post('/users/authenticate', [UserController::class, 'authenticate']);
+
 
 // example:
 // Route::get('/hello', function(){
